Support os_cacert
Currently os_cacert is not respected for the underlying Ironic
and Heat clients because it is not passed on to those clients.

This does currently require accessing a private member of the
self._instance object, but in looking at the upstream ironic plugin
I see it is also doing that (and the code there seems to
automatically handle the cacert option), so I think we can revisit
the question once the upstream plugin merges.  If it's considered
acceptable there then it probably is here too.

Also, the upstream plugin is using a different method of creating
the client, and at some point we should converge the two, so this
code may go away in the future anyway.

Change-Id: Ic122b0eedc6ed99bd11b2d0c347e7e9740e6dba6
(cherry picked from commit 10f88ab692ff1713f2e0315066174fe9bec7f717)
Related: rhbz#1230890